requirements.txt
accelerate==1.2.0
gsplat==1.0.0
open3d==0.18.0
tqdm==4.64.1
omegaconf==2.2.3
opencv-python==4.9.0.80
opencv-contrib-python==4.8.0.74
plyfile==1.0.3
timm==0.9.2
wandb==0.17.4
ftfy==6.2.0
diffdist==0.1
diffusers==0.31.0
einops==0.4.1
imageio==2.34.1
imageio-ffmpeg==0.5.1
transformers==4.47.0
torchsde
huggingface-hub==0.26.3
torchmetrics=1.3.2
natten==0.14.6
peft